Add `shr_RandNum_mod`.
`shr_RandNum_mod` is the Fortran interface for a collection of random
number generator codes, collected and optimized by Chris Kerr.

This commit simply adds the files and a benchmark/test, with some
modifications to the code and directory structure to simplify the build
and for portability. The main deviations from the original are:

 1) The input to create a random number generator is now a single 2D
    seed array, rather than a set of optional arguments for different
    RNG methods.

 2) The original test file is thrown out, and the benchmark/test program
    is instead built through the CMake unit test system.

This code still needs to be made more portable across compilers, and it
needs additional file-specific compiler flags in order to get the
good performance seen in Chris's version.

Testing:

    Test suite: Unit and benchmark tests run, using GCC only.
    Baseline: N/A
    Namelist changes: None.
    Status: bit for bit

Code reviews:

    Informal - Integration/review by Sean and conversation with Chris.
               Ongoing.