Change some bad variables in config_compilers.
PIO_TYPENAME was being set in the gnu/yellowstone section. This setting
is removed. It's not clear if this setting had any effect in the first
place; regardless, PIO_TYPENAME should be set in the environment, not
the config_compilers.xml file.

ADD_CPPFLAGS was being set for AIX machines. Unless there is a bug
somewhere, this should be equivalent to ADD_CPPDEFS, which is the
preferred variable for adding preprocessor flags. The name on this
element has therefore been changed.

This commit is not tested. It appears extremely unlikely that this will
affect the build on any supported machine, however.