Put in ambiguity check to config_build parsing.
There are two cases where we could have an ambiguous config_build
setting in the XML.

Firstly, we could have something like this:

<MPI_PATH MPILIB="mpich">/some/path</MPI_PATH>
<MPI_PATH MPILIB="mpich">/some/other/path</MPI_PATH>

There's obviously an inconsistency when you set two values for the same
variable in the same situation, so we want the user to know about it.

Secondly, we could have a situation like this:

<MPI_PATH MPILIB="mpich">/some/mpich/path</MPI_PATH>
<MPI_PATH DEBUG="TRUE">/some/openmpi/debug/path</MPI_PATH>

In this case, if you have a debug build that uses MPICH, it's not clear
what should happen. Should the DEBUG library override the MPILIB
setting? Or should the MPILIB setting always force use of the MPICH
library, regardless of DEBUG? Or did the user forget to specify some
third library, an MPICH build specifically configured for debugging?
We want to throw an error in this case as well to force the user to be
more specific.

This commit institutes a check for these two situations in perl5lib, in
Build::MacroMatchList, in the form of a consistency check called
"ambiguity_check" that is called when converting the list to a
Build::MacroMatchTree.