Add CMake output to MacroMaker.
This is done primarily through adding a new MacroCMakeWriter class. This
change also required adding some new prepend/append information for
nodes. The reason for this is that, unlike in Makefiles, in CMake it's
not possible to execute a command inline while setting a variable; you
have to execute the command in a separate statement beforehand, which
assigns to a temporary a variable, then reference that variable inline.