Fix usage of multiple <shell> tags with CMake.
Due to the fact that CMake needs a temporary variable to hold the output
of executed processes, using multiple <shell> tags was not working due
to a name conflict with this variable. This was solved by modifying
MacroCMakeWriter to keep a counter to be incorporated into the variable
name, and incrementing the counter every time a <shell> tag is used.