Fix settings for mpi-serial cases on yellowstone.
Most of the yellowstone-specific Intel settings were set only for
Macros files that were created with `MPILIB="mpich2"`. By changing the
locations where the `MPILIB` attribute is used in `config_compilers`,
this commit makes the settings for mpi-serial cases the same as for
MPICH2 cases, and avoids causing the Macros file to be specific to a
specific MPI library.

Testing:

    None! This might be (in fact seems like it should be) answer-
    changing for mpi-serial builds on yellowstone.