Reconcile `datm` namelist defaults and definition.
Several items in the namelist defaults file for DATM were not actually
listed in the definition file, or were listed under a different name.
This commit fixes this using changes to both files.

`DATM_MODE` and `DATM_PRESAERO` are also converted to lowercase in the
definition file, since keeping all variable names lowercase makes it
easier to process the file.